+++ app/src/main/java/com/travelguide/fragments/ProfileFragment.java
@@ -24,7 +24,7 @@ import com.travelguide.adapters.TripPlanAdapter;
 import com.travelguide.decorations.VerticalSpaceItemDecoration;
 import com.travelguide.helpers.DeviceDimensionsHelper;
 import com.travelguide.models.TripPlan;
-import com.travelguide.utils.ItemClickSupport;
+import com.travelguide.helpers.ItemClickSupport;
 
 import java.util.ArrayList;
 import java.util.List;
